added SPF support.
NSD supports more draft-RR's, so one cannot harm.

I had to stretch-up the array holding all the rrtypes,
because SPF has been assigned type code 99 and we
stopped at 48 before that. I don't know what performance
impact that has on zonec.

On the other hand one day in the future we would be
forced to extend this array anyhow for newer RR types.



git-svn-id: http://www.nlnetlabs.nl/svn/nsd@1615 a26ef69c-88ff-0310-839f-98b793d9c207